Move Authenticatrs to `arbeitszeit_web.www`-folder
The MemberAuthenticator and CompanyAuthenticator seem to be www-specific, that's why they were moved to the `www` directory.
